# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Example of clean architecture in Go.",
        "title": "Go E-Presciprion Clean API",
        "contact": {},
        "version": "1.0"
    },
    "host": "http://localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/medicine-details": {
            "get": {
                "description": "Retrieve all medicine detail records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine-detail"
                ],
                "summary": "Get all medicine details",
                "operationId": "get-all-medicine-details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessMedicineDetailList"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new medicine detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine-detail"
                ],
                "summary": "Create medicine detail",
                "operationId": "create-medicine-detail",
                "parameters": [
                    {
                        "description": "MedicineDetail input",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MedicineDetail"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessMedicineDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/medicine-details/{id}": {
            "get": {
                "description": "Retrieve a single medicine detail by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine-detail"
                ],
                "summary": "Get medicine details by ID",
                "operationId": "get-medicine-detail-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MedicineDetail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessMedicineDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a medicine detail by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine-detail"
                ],
                "summary": "Delete medicine detail",
                "operationId": "delete-medicine-detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MedicineDetail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessString"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a medicine detail by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine-detail"
                ],
                "summary": "Update medicinedetail",
                "operationId": "update-medicine-detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MedicineDetail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "MedicineDetail input",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MedicineDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessMedicineDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/medicines": {
            "get": {
                "description": "Retrieve all medicine records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine"
                ],
                "summary": "Get all medicines",
                "operationId": "get-all-medicines",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessMedicineList"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new medicine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine"
                ],
                "summary": "Create Medicine",
                "operationId": "create-medicine",
                "parameters": [
                    {
                        "description": "Medicine input",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Medicine"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessMedicine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/medicines/{id}": {
            "get": {
                "description": "Retrieve a single medicine by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine"
                ],
                "summary": "Get medicine by ID",
                "operationId": "get-by-id-medicine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessMedicine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a medicine by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine"
                ],
                "summary": "Delete medicine",
                "operationId": "delete-medicine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessString"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update an existing medicine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine"
                ],
                "summary": "Update medicine",
                "operationId": "update-medicine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medicine input",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Medicine"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessMedicine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Medicine": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "1789dfb3-e060-433a-9db4-cb93b42768d8"
                },
                "name": {
                    "type": "string",
                    "example": "Panadol"
                },
                "price": {
                    "type": "integer",
                    "example": 10000
                },
                "quantity": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "entity.MedicineDetail": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "For pain relief"
                },
                "id": {
                    "type": "string",
                    "example": "1789dfb3-e060-433a-9db4-cb93b42768d8"
                },
                "medicine_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                },
                "signa_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "transaction_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
...
        "request.Medicine": {
            "type": "object",
            "required": [
                "name",
                "price",
                "quantity"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Panadol"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 10000
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 100
                }
            }
        },
        "request.MedicineDetail": {
            "type": "object",
            "required": [
                "quantity"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "For pain relief"
                },
                "medicine_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 2
                },
                "signa_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "response.SuccessMedicine": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.Medicine"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.SuccessMedicineDetail": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.MedicineDetail"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.SuccessMedicineDetailList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.MedicineDetail"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.SuccessMedicineList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Medicine"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.SuccessString": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "deleted"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    }
}